Fixed the API JSON data keys
We were coming from the data grid example, and we left the headers names in
the data for no good reason. It makes sense to think of this as a much better
example, and use decent keys for the data, and that's what we did. Simply
change the way we're picking it off in the JavaScript, and the server.clj and
we're all good. Update the docs, and it's a trifecta.