Added fading on the status block
We don't need the status block to hang around forever - it gives the wrong
impression. Better to let it show, and then disappear. So we needed to correct
the .replaceWIth() with .html() as that's how to replace the *contents* and
not the whole thing. That made it possible, and then we just needed to add
in the .show() and then .fadeOut(), and we're good to go. At the same time,
the 'Undo' button forces the .hide() to make sure it's not there any more.