Added the Google Identity service to the code
Jeff thought it'd be nice to have an OAuth2 security system in the code, so
he found the Google Identity service, and I plugged it in. The code is
pretty easy, just pick off the headers with the auth token, and then check
it with Google, and then check the email with a list we know to be good. It's
simple, but it works, and with Google, it works pretty well.